Merge test scenarios in one to reduce execution time
Opening the web page(sahibinden.com) takes too much time. When there was a scenario for each element in the header, it took 26 seconds for tests to finish executing. After reducing the test scenarios to 1 which reduced the number of opening the web page it took 3 seconds for tests to complete.